requirements.txt
pandas>=1.3.0
numpy>=1.21.0
streamlit>=1.22.0
scikit-learn>=0.24.2
scikit-surprise>=1.1.1
plotly>=5.3.0
altair>=4.2.0
python-dotenv>=0.19.0
typing-extensions>=4.0.0
requests>=2.26.0 